# openapi.yml
openapi: 3.0.3
info:
  title: Hospital Management System API
  description: |
    A comprehensive web-based Hospital Management System built with Flask.

    This API provides role-based access control for three user types:
    - **Admin**: Manages doctors, patients, and appointments
    - **Doctor**: Manages appointments, patient records, and availability
    - **Patient**: Books appointments and accesses treatment history

    ## Authentication
    The API uses session-based authentication with Flask-Login.
    Login is required for most endpoints with role-based access control.

    ## Security
    - Bcrypt password hashing
    - CSRF protection on all forms
    - SQL injection prevention via ORM
    - Role-based access control (RBAC)
  version: 0.5.0
  contact:
    name: IIT Madras Student
    email: admin@hospital.com
  license:
    name: Educational Use Only
    url: https://www.iitm.ac.in
servers:
  - url: http://127.0.0.1:5000
    description: Development server
  - url: http://localhost:5000
    description: Local server
...
paths:
...
  /admin/dashboard:
    get:
      tags:
        - Admin
      summary: Admin dashboard
      description: Display admin dashboard with statistics
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Dashboard displayed with statistics
          content:
            text/html:
              schema:
                type: string
        '403':
          description: Access denied - Admin role required
  /admin/doctors:
    get:
      tags:
        - Admin
      summary: List all doctors
      description: Display all doctors with search functionality
      security:
        - sessionAuth: []
      parameters:
        - name: search
          in: query
          description: Search term for doctor name or specialization
          schema:
            type: string
      responses:
        '200':
          description: List of doctors
  /admin/doctors/add:
    get:
      tags:
        - Admin
      summary: Add doctor form
      description: Display form to add new doctor
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Add doctor form displayed
    post:
      tags:
        - Admin
      summary: Create new doctor
      description: Add new doctor to the system
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/DoctorCreate'
      responses:
        '302':
          description: Doctor created successfully
        '400':
          description: Validation error
  /admin/doctors/edit/{id}:
    get:
      tags:
        - Admin
      summary: Edit doctor form
      description: Display form to edit doctor details
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Edit doctor form displayed
        '404':
          description: Doctor not found
    post:
      tags:
        - Admin
      summary: Update doctor
      description: Update doctor information
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/DoctorUpdate'
      responses:
        '302':
          description: Doctor updated successfully
        '404':
          description: Doctor not found
  /admin/doctors/delete/{id}:
    post:
      tags:
        - Admin
      summary: Delete doctor
      description: Remove doctor from the system
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '302':
          description: Doctor deleted successfully
        '404':
          description: Doctor not found
  /admin/patients:
    get:
      tags:
        - Admin
      summary: List all patients
      description: Display all patients with search functionality
      security:
        - sessionAuth: []
      parameters:
        - name: search
          in: query
          description: Search term for patient name or email
          schema:
            type: string
      responses:
        '200':
          description: List of patients
  /admin/patients/edit/{id}:
    get:
      tags:
        - Admin
      summary: Edit patient form
      description: Display form to edit patient details
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Edit patient form displayed
        '404':
          description: Patient not found
    post:
      tags:
        - Admin
      summary: Update patient
      description: Update patient information
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatientUpdate'
      responses:
        '302':
          description: Patient updated successfully
        '404':
          description: Patient not found
  /admin/appointments:
    get:
      tags:
        - Admin
      summary: View all appointments
      description: Display all appointments in the system
      security:
        - sessionAuth: []
      parameters:
        - name: status
          in: query
          description: Filter by appointment status
          schema:
            type: string
            enum: [Booked, Completed, Cancelled]
      responses:
        '200':
          description: List of appointments
  /admin/specializations:
    get:
      tags:
        - Admin
      summary: Manage specializations
      description: View and manage medical specializations
      security:
        - sessionAuth: []
      responses:
        '200':
          description: List of specializations

  # ==================== DOCTOR ROUTES ====================
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session
      description: Flask session-based authentication
  schemas:
...
    DoctorCreate:
      type: object
      required:
        - username
        - email
        - password
        - name
        - specialization_id
        - license_number
        - experience_years
        - consultation_fee
      properties:
        username:
          type: string
          example: dr.smith
        email:
          type: string
          format: email
          example: dr.smith@hospital.com
        password:
          type: string
          example: doctor123
        name:
          type: string
          example: Dr. John Smith
        specialization_id:
          type: integer
          example: 1
        license_number:
          type: string
          example: MED12345
        experience_years:
          type: integer
          minimum: 0
          example: 10
        consultation_fee:
          type: number
          format: float
          example: 500.00
        phone:
          type: string
          example: '+91 9876543210'
    DoctorUpdate:
      type: object
      properties:
        name:
          type: string
          example: Dr. John Smith
        specialization_id:
          type: integer
          example: 1
        license_number:
          type: string
          example: MED12345
        experience_years:
          type: integer
          example: 10
        consultation_fee:
          type: number
          format: float
          example: 500.00
        phone:
          type: string
          example: '+91 9876543210'
    PatientUpdate:
      type: object
      properties:
        name:
          type: string
          example: John Doe
        age:
          type: integer
          example: 35
        gender:
          type: string
          enum: [Male, Female, Other]
        phone:
          type: string
          example: '+91 9876543210'
        address:
          type: string
          example: '123 Main St, Mumbai'
        is_active:
          type: boolean
          example: true
...
security:
  - sessionAuth: []